# packages/aio-commerce-lib-app/generated/openapi.gen.json
{
  "info": {
    "title": "AIO Commerce SDK API",
    "version": "1.3.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/scope-tree/": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scopes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "level": {
                            "type": "string"
                          },
                          "is_editable": {
                            "type": "boolean"
                          },
                          "is_final": {
                            "type": "boolean"
                          },
                          "is_removable": {
                            "type": "boolean"
                          },
                          "commerce_id": {
                            "type": "number"
                          },
                          "children": {
                            "$ref": "#/components/schemas/Ref9"
                          }
                        },
                        "required": [
                          "id",
                          "code",
                          "label",
                          "level",
                          "is_editable",
                          "is_final",
                          "is_removable"
                        ]
                      }
                    }
                  },
                  "required": [
                    "scopes"
                  ]
                }
              }
            },
            "description": "Success"
          },
          "203": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scopes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "level": {
                            "type": "string"
                          },
                          "is_editable": {
                            "type": "boolean"
                          },
                          "is_final": {
                            "type": "boolean"
                          },
                          "is_removable": {
                            "type": "boolean"
                          },
                          "commerce_id": {
                            "type": "number"
                          },
                          "children": {
                            "$ref": "#/components/schemas/Ref10"
                          }
                        },
                        "required": [
                          "id",
                          "code",
                          "label",
                          "level",
                          "is_editable",
                          "is_final",
                          "is_removable"
                        ]
                      }
                    }
                  },
                  "required": [
                    "scopes"
                  ],
                  "description": "Cached scope tree"
                }
              }
            },
            "description": "Cached scope tree"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "scopes": {
                    "type": "array",
                    "items": {}
                  }
                },
                "required": [
                  "scopes"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "result": {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string"
                        },
                        "timestamp": {
                          "type": "string"
                        },
                        "scopes": {
                          "type": "array",
                          "items": {
                            "type": "object",
                            "properties": {
                              "id": {
                                "type": "string"
                              },
                              "code": {
                                "type": "string"
                              },
                              "label": {
                                "type": "string"
                              },
                              "level": {
                                "type": "string"
                              },
                              "is_editable": {
                                "type": "boolean"
                              },
                              "is_final": {
                                "type": "boolean"
                              },
                              "children": {
                                "$ref": "#/components/schemas/Ref11"
                              }
                            },
                            "required": [
                              "id",
                              "code",
                              "label",
                              "level",
                              "is_editable",
                              "is_final"
                            ]
                          }
                        }
                      },
                      "required": [
                        "message",
                        "timestamp",
                        "scopes"
                      ]
                    }
                  },
                  "required": [
                    "result"
                  ]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    },
    "/scope-tree/commerce": {
      "post": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "commerceBaseUrl": {
                    "type": "string",
                    "minLength": 1
                  },
                  "commerceEnv": {
                    "type": "string",
                    "minLength": 1
                  }
                },
                "required": [
                  "commerceBaseUrl"
                ]
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scopes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "level": {
                            "type": "string"
                          },
                          "is_editable": {
                            "type": "boolean"
                          },
                          "is_final": {
                            "type": "boolean"
                          },
                          "is_removable": {
                            "type": "boolean"
                          },
                          "commerce_id": {
                            "type": "number"
                          },
                          "children": {
                            "$ref": "#/components/schemas/Ref12"
                          }
                        },
                        "required": [
                          "id",
                          "code",
                          "label",
                          "level",
                          "is_editable",
                          "is_final",
                          "is_removable"
                        ]
                      }
                    },
                    "synced": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "scopes",
                    "synced"
                  ]
                }
              }
            },
            "description": "Success"
          },
          "203": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "scopes": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "id": {
                            "type": "string"
                          },
                          "code": {
                            "type": "string"
                          },
                          "label": {
                            "type": "string"
                          },
                          "level": {
                            "type": "string"
                          },
                          "is_editable": {
                            "type": "boolean"
                          },
                          "is_final": {
                            "type": "boolean"
                          },
                          "is_removable": {
                            "type": "boolean"
                          },
                          "commerce_id": {
                            "type": "number"
                          },
                          "children": {
                            "$ref": "#/components/schemas/Ref13"
                          }
                        },
                        "required": [
                          "id",
                          "code",
                          "label",
                          "level",
                          "is_editable",
                          "is_final",
                          "is_removable"
                        ]
                      }
                    },
                    "synced": {
                      "type": "boolean"
                    }
                  },
                  "required": [
                    "scopes",
                    "synced"
                  ],
                  "description": "Cached scope tree"
                }
              }
            },
            "description": "Cached scope tree"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      },
      "delete": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string"
                    }
                  },
                  "required": [
                    "message"
                  ]
                }
              }
            },
            "description": "Success"
          },
          "400": {
            "description": "Bad request"
          },
          "404": {
            "description": "Not found"
          },
          "405": {
            "description": "Method not allowed"
          },
          "500": {
            "description": "Internal server error"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Ref9": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "code": {
              "type": "string"
            },
            "label": {
              "type": "string"
            },
            "level": {
              "type": "string"
            },
            "is_editable": {
              "type": "boolean"
            },
            "is_final": {
              "type": "boolean"
            },
            "is_removable": {
              "type": "boolean"
            },
            "commerce_id": {
              "type": "number"
            },
            "children": {
              "$ref": "#/components/schemas/Ref9"
            }
          },
          "required": [
            "id",
            "code",
            "label",
            "level",
            "is_editable",
            "is_final",
            "is_removable"
          ]
        }
      },
      "Ref10": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "code": {
              "type": "string"
            },
            "label": {
              "type": "string"
            },
            "level": {
              "type": "string"
            },
            "is_editable": {
              "type": "boolean"
            },
            "is_final": {
              "type": "boolean"
            },
            "is_removable": {
              "type": "boolean"
            },
            "commerce_id": {
              "type": "number"
            },
            "children": {
              "$ref": "#/components/schemas/Ref10"
            }
          },
          "required": [
            "id",
            "code",
            "label",
            "level",
            "is_editable",
            "is_final",
            "is_removable"
          ]
        }
      },
      "Ref11": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "code": {
              "type": "string"
            },
            "label": {
              "type": "string"
            },
            "level": {
              "type": "string"
            },
            "is_editable": {
              "type": "boolean"
            },
            "is_final": {
              "type": "boolean"
            },
            "children": {
              "$ref": "#/components/schemas/Ref11"
            }
          },
          "required": [
            "id",
            "code",
            "label",
            "level",
            "is_editable",
            "is_final"
          ]
        }
      },
      "Ref12": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "code": {
              "type": "string"
            },
            "label": {
              "type": "string"
            },
            "level": {
              "type": "string"
            },
            "is_editable": {
              "type": "boolean"
            },
            "is_final": {
              "type": "boolean"
            },
            "is_removable": {
              "type": "boolean"
            },
            "commerce_id": {
              "type": "number"
            },
            "children": {
              "$ref": "#/components/schemas/Ref12"
            }
          },
          "required": [
            "id",
            "code",
            "label",
            "level",
            "is_editable",
            "is_final",
            "is_removable"
          ]
        }
      },
      "Ref13": {
        "type": "array",
        "items": {
          "type": "object",
          "properties": {
            "id": {
              "type": "string"
            },
            "code": {
              "type": "string"
            },
            "label": {
              "type": "string"
            },
            "level": {
              "type": "string"
            },
            "is_editable": {
              "type": "boolean"
            },
            "is_final": {
              "type": "boolean"
            },
            "is_removable": {
              "type": "boolean"
            },
            "commerce_id": {
              "type": "number"
            },
            "children": {
              "$ref": "#/components/schemas/Ref13"
            }
          },
          "required": [
            "id",
            "code",
            "label",
            "level",
            "is_editable",
            "is_final",
            "is_removable"
          ]
        }
      }
    }
  }
}